requirements.txt
astroid==2.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-stubs==0.4.1
entrypoints==0.3
httpie==1.0.2
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.3
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
numpy==1.15.4
pandocfilters==1.4.2
parso==0.3.4
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.5.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pycryptodome==3.7.2
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
retrying==1.3.3
Send2Trash==1.5.0
simple-crypt==4.1.7
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.0
